chore: refactor react components
