+++ packages/client/src/mud/components/MoveCardComponent.ts
@@ -1,9 +1,8 @@
-import { defineComponent, Type, World } from "@latticexyz/recs";
+import { defineComponent, Type } from "@latticexyz/recs";
+import { world } from "../world";
 
-export function defineMoveCardComponent(world: World) {
-  return defineComponent(
-    world,
-    { distance: Type.Number, direction: Type.Number, rotation: Type.Number },
-    { id: "MoveCard", metadata: { contractId: "ds.component.MoveCard" } }
-  );
-}
+export const MoveCardComponent = defineComponent(
+  world,
+  { distance: Type.Number, direction: Type.Number, rotation: Type.Number },
+  { id: "MoveCard", metadata: { contractId: "ds.component.MoveCard" } }
+);
