+++ packages/client/src/mud/components/index.ts
@@ -1,1 +1,1 @@
-export { defineLoadingStateComponent } from "./LoadingStateComponent";
+export { clientComponents, components } from "./components";
