--- packages/client/src/mud/constants.ts
@@ -1,1 +1,0 @@
-export const GAS_LIMIT = 30_000_000;
