--- packages/client/src/mud/index.ts
@@ -1,2 +1,0 @@
-export { createNetworkLayer } from "../../setupMUD";
-export type { NetworkLayer } from "./types";
