--- packages/client/src/mud/setup/index.ts
@@ -1,1 +1,0 @@
-export { setupDevSystems } from "../setupDevSystems";
