+++ packages/client/src/react/HoveredShip.tsx
@@ -1,6 +1,6 @@
 import { useComponentValue } from "@latticexyz/react";
 import { EntityIndex } from "@latticexyz/recs";
-import { useMUD } from "../MUDContext";
+import { useMUD } from "../mud/providers/MUDProvider";
 import { Cell } from "./components/Cell";
 import { ShipCard } from "./components/ShipStatus/ShipCard";
 import { Container, InternalContainer } from "./styles/global";
