+++ packages/client/src/react/components/DamageChance.tsx
@@ -2,7 +2,7 @@ import { useComponentValue, useObservableValue } from "@latticexyz/react";
 import { EntityIndex } from "@latticexyz/recs";
 import { Coord } from "@latticexyz/utils";
 import styled from "styled-components";
-import { useMUD } from "../../MUDContext";
+import { useMUD } from "../../mud/providers/MUDProvider";
 import { colors, Container } from "../styles/global";
 
 type ShipData = {
