+++ packages/client/src/react/components/ShipStatus/ConfirmButtons.tsx
@@ -2,8 +2,8 @@ import { useComponentValue, useObservableValue } from "@latticexyz/react";
 import { getComponentEntities, getComponentValue, getComponentValueStrict, Has, runQuery } from "@latticexyz/recs";
 import { merge } from "rxjs";
 import styled from "styled-components";
-import { useMUD } from "../../../MUDContext";
-import { usePlayer } from "../../../PlayerContext";
+import { useMUD } from "../../../mud/providers/MUDProvider";
+import { usePlayer } from "../../../mud/providers/PlayerProvider";
 import { ActionType, Phase } from "../../../types";
 import { colors, Success } from "../../styles/global";
 import { ActionButtons } from "./ActionButtons";
