+++ packages/client/src/react/components/ShipStatus/YourShips.tsx
@@ -2,9 +2,9 @@ import { useComponentValue, useObservableValue } from "@latticexyz/react";
 import { EntityIndex, getComponentValue, Has, HasValue, runQuery, setComponent } from "@latticexyz/recs";
 import { merge } from "rxjs";
 import styled from "styled-components";
+import { useMUD } from "../../../mud/providers/MUDProvider";
+import { usePlayer } from "../../../mud/providers/PlayerProvider";
 import { world } from "../../../mud/world";
-import { useMUD } from "../../../MUDContext";
-import { usePlayer } from "../../../PlayerContext";
 import { ModalType, Phase } from "../../../types";
 import { Button, colors, Container } from "../../styles/global";
 import { Cell } from "./../Cell";
