+++ packages/client/src/react/components/TurnTimer.tsx
@@ -1,8 +1,8 @@
 import { useObservableValue } from "@latticexyz/react";
 import { getComponentValue } from "@latticexyz/recs";
 import styled, { keyframes } from "styled-components";
-import { useMUD } from "../../MUDContext";
-import { usePlayer } from "../../PlayerContext";
+import { useMUD } from "../../mud/providers/MUDProvider";
+import { usePlayer } from "../../mud/providers/PlayerProvider";
 import { Category } from "../../sound";
 import { Phase } from "../../types";
 import { colors } from "../styles/global";
