+++ packages/contracts/package.json
@@ -9,7 +9,7 @@
     "devnode": "mud devnode",
     "test": "mud test",
     "types": "mud types",
-    "dev": "yarn prepare && mud deploy-contracts --dev --watch --openUrl http://localhost:3000?dev=true",
+    "dev": "mud deploy-contracts --dev --watch --openUrl http://localhost:3000?dev=true",
     "deploy": "mud deploy-contracts",
     "deploy:prod": "mud deploy --deployerPrivateKey 0x26e86e45f6fc45ec6e2ecd128cec80fa1d1505e5507dcd2ae58c3130a7a97b48 --chainSpec ./lattice-chain-spec.json --deployClient --netlifySlug mud --netlifyPersonalToken q49-wyrS2OGAMk60KsCyfyQGD0QgCcjfqKmtiu8RtQc",
     "lint": "yarn prettier && yarn solhint",
